chore: tidy up file and break down into ts files
